Close #321 - Move loggerf.cats.instances.logF to loggerf.instances.cats
