+++ modules/logger-f-cats-effect3/shared/src/test/scala/loggerf/cats/syntaxSpec.scala
@@ -80,7 +80,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO].completeThen { _ =>
       logger ==== expected
@@ -125,7 +125,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](logMsg).completeThen { _ =>
       logger ==== expected
@@ -169,7 +169,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](logMsg).completeThen { _ =>
       logger ==== expected
@@ -214,7 +214,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](logMsg).completeThen { _ =>
       logger ==== expected
@@ -261,7 +261,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](eab).completeThen { _ =>
       logger ==== expected
@@ -308,7 +308,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](eab).completeThen { _ =>
       logger ==== expected
@@ -355,7 +355,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](eab).completeThen { _ =>
       logger ==== expected
@@ -399,7 +399,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](logMsg).completeThen { _ =>
       logger ==== expected
@@ -443,7 +443,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](logMsg).completeThen { _ =>
       logger ==== expected
@@ -487,7 +487,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](logMsg).completeThen { _ =>
       logger ==== expected
@@ -534,7 +534,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](eab).completeThen { _ =>
       logger ==== expected
@@ -581,7 +581,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](eab).completeThen { _ =>
       logger ==== expected
@@ -628,7 +628,7 @@ object syntaxSpec extends Properties {
 
     import CatsEffectRunner._
     import effectie.ce3.fx.ioFx
-    import loggerf.cats.instances.logF
+    import loggerf.instances.cats.logF
     implicit val ticket: Ticker = Ticker(TestContext())
     runLog[IO](eab).completeThen { _ =>
       logger ==== expected
@@ -665,7 +665,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO].completeThen { _ =>
         logger ==== expected
@@ -710,7 +710,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](logMsg).completeThen { _ =>
         logger ==== expected
@@ -754,7 +754,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](logMsg).completeThen { _ =>
         logger ==== expected
@@ -799,7 +799,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](logMsg).completeThen { _ =>
         logger ==== expected
@@ -846,7 +846,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](eab).completeThen { _ =>
         logger ==== expected
@@ -893,7 +893,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](eab).completeThen { _ =>
         logger ==== expected
@@ -940,7 +940,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](eab).completeThen { _ =>
         logger ==== expected
@@ -984,7 +984,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](logMsg).completeThen { _ =>
         logger ==== expected
@@ -1028,7 +1028,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](logMsg).completeThen { _ =>
         logger ==== expected
@@ -1072,7 +1072,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](logMsg).completeThen { _ =>
         logger ==== expected
@@ -1119,7 +1119,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](eab).completeThen { _ =>
         logger ==== expected
@@ -1166,7 +1166,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](eab).completeThen { _ =>
         logger ==== expected
@@ -1213,7 +1213,7 @@ object syntaxSpec extends Properties {
 
       import CatsEffectRunner._
       import effectie.ce3.fx.ioFx
-      import loggerf.cats.instances.logF
+      import loggerf.instances.cats.logF
       implicit val ticket: Ticker = Ticker(TestContext())
       runLog[IO](eab).completeThen { _ =>
         logger ==== expected
